Cleanup to avoid to check for trivial bugs - bug is related to only tracking one perf_region ie region 0

+++ bench/src/benchmark_names.f90
@@ -1,8 +1,8 @@
 MODULE benchmark_names
     implicit none
     ! see https://pages.mtu.edu/~shene/COURSES/cs201/NOTES/chap02/param.html
-    integer, parameter :: TEST_BENCH_0 = 0
-    integer, parameter :: TEST_BENCH_1 = 1
-    integer, parameter :: TEST_BENCH_2D_JI = 2
-    integer, parameter :: TEST_BENCH_2D_IJ = 3
+    integer(KIND=8), parameter :: TEST_BENCH_0 = 0
+    integer(KIND=8), parameter :: TEST_BENCH_1 = 1
+    integer(KIND=8), parameter :: TEST_BENCH_2D_JI = 2
+    integer(KIND=8), parameter :: TEST_BENCH_2D_IJ = 3
 END MODULE benchmark_names
